requirements.txt
asgiref==3.4.1
Django==3.2
django-render-partial==0.4
Pillow==8.4.0
psycopg2-binary==2.9.2
pytz==2021.3
sorl-thumbnail==12.7.0
sqlparse==0.4.2
suds-py3==1.4.5.0
